+++ contracts/member/interfaces/IRewards.sol
@@ -3,8 +3,12 @@ pragma solidity ^0.8.4;
 
 interface IRewards {
     function rewardToken() external returns (address);
-    function updateShares(uint deposits, address token) external;
-    function getRewardsIssued() external view returns (uint);
-    function issueRewards(address recipient, uint amount) external;
-    function updateMultiplier(uint newMultiplier) external;
+
+    function updateShares(uint256 deposits, address token) external;
+
+    function getRewardsIssued() external view returns (uint256);
+
+    function issueRewards(address recipient, uint256 amount) external;
+
+    function updateMultiplier(uint256 newMultiplier) external;
 }
