+++ contracts/token/Vesting.sol
@@ -6,19 +6,18 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "./ModlMint.sol";
 
 contract Vesting is AccessControl {
-
     ModlMint minter;
 
-    uint startedAt;
+    uint256 startedAt;
 
     struct VestingState {
-        uint totalAmount;
-        uint startTime;
-        uint endTime;
-        uint claimed;
+        uint256 totalAmount;
+        uint256 startTime;
+        uint256 endTime;
+        uint256 claimed;
     }
 
-    mapping (address => VestingState) vestingStates;
+    mapping(address => VestingState) vestingStates;
 
     modifier started() {
         require(startedAt > 0);
@@ -35,14 +34,14 @@ contract Vesting is AccessControl {
     }
 
     function claim(address addr) public {
-        require(msg.sender == addr || hasRole(DEFAULT_ADMIN_ROLE, msg.sender), "Cannot claim for other people.");
+        require(
+            msg.sender == addr || hasRole(DEFAULT_ADMIN_ROLE, msg.sender),
+            "Cannot claim for other people."
+        );
         minter.mint(addr, claimable(addr));
-    } 
+    }
 
-    function claimable(address addr) public returns (uint) {
-
-        if (block.timestamp > vestingStates[addr].endTime) {
-
-        }
+    function claimable(address addr) public returns (uint256) {
+        if (block.timestamp > vestingStates[addr].endTime) {}
     }
 }
